Possible (very simple) optimization to OrderedSetX
Both JDKGDXDS' OrderedSet (which is roughly equivalent to the Merry or libGDX 1.9.11 OrderedSet) and OrderedSetX slow down a lot compared to an ObjectSet or ObjectSetX on composite benchmarks like add-or-remove. Profiling indicates that there's a lot of unnecessary branching in OrderedSet(X)'s add() and remove() methods. LET'S SEE HOW THIS GOES!